+++ NEWS.md
@@ -6,6 +6,7 @@
 1. Fix #277 - Add function `ts_auto_arima()` boiler plate function.
 2. Fix #284 - Add functions `color_blind()` `ts_scale_fill_colorblind()` and
 `ts_scale_color_colorblind()`
+3. Fix #278 - Add function `ts_auto_smooth_es()`
 
 ## Minor Fixes and Improvements
 1. Fix #275 - Correct `ts_auto_arima_xgboots()` when `.tune` is FALSE.